+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## future
 ### New features
-- Mpris support. On linux, OS will recognize uamp as running media player.
+- Mpris support. On linux, OS will recognize uamp as running media player. This
+  is enabled by default and may be disabled with a new option `system_player`.
 
 ### Fixes
 - Fix issue where some images would accidentally use default background color